Enemy and enemy types
The enemy class contains core variables such as health, damage, XP/gold rewards and enemy types. The enemy type instance and the toString method will allow the player to understand what type of enemy they are facing.

Depending on which type the enemy is will decide on what health, damage and rewards they have which is currently initialized in the constructor. This will be useful when the dungeon class is created as it will allow for specific enemies to be spawned depending on the player's level.