Dungeon class creation
This class is a representation of a visitable location in which the player can encounter different types of enemies depending on their level.

Whilst the player doesn't get a visual notification (unless it is empty), the dungeon will always contain a different amount of enemies, to allow for randomisation and difficulty spikes.