Town interactions and tavern
Implemented interactions within the town such as the guild and blacksmith which currently only print out information until the purchasing functionality is setup.

Additionally the alchemist location has been changed to the tavern to make more sense gameplay wise as it's primary function was to only provide the player a way to fully restore their health without the house.