Partial interaction and dragon enemy
Added interactivity for choices which don't require the purchasing functionality and only transitions to locations or repeats the location transition if the users input is invalid.

Also included a dragon as an enemy type which only spawns once the player enters the castle and is level 5 (currently max level). This is to include a final boss for the game rather than re-use existing ones that can be faced in the dungeon, in which the dungeon class now has a getRandomEnemy method to allow for randomisation so that each enemy encounter could be either against a goblin or an orc.