Purchase implementation
Added the core purchasing functionality which now allows the player to buy assets/upgrades.

This is done through the usage of an Enum that contains all of the assets/upgrades the player can purchase which is then passed through as an argument in the purchaseAsset function that checks if the player has enough gold which, if true, increases the players stats (e.g. health increase) or allows them to restore their health for free without having to spend gold on a tavern room.

Also added in system("cls") commands to allow for the terminal to declutter previous std::cout commands and see more important information easily.