Combat interaction and location loops
Added the combat function which has a multipurpose usage instead of creating separate functions for facing the final boss and enemies in the dungeon, as well as a gameEnd function that prints out an appropriate game over message whether the player has beaten the game or been defeated by the enemy.

To avoid the program from terminating prematurely, the implementation of the while loops alongside the in<location name> Booleans ensures that after the player has made a choice (e.g. purchased a upgrade from the blacksmith) the program allows the player to make another until they close the program or by other means.