Combat change & interaction fixes
Due to the nature of the final boss and the dungeon having different outcomes when the enemy is defeated (the final boss ends the game, whereas the dungeon allows the player to face another enemy), the combat function instead is separated into location specific functions (dungeonCombat and finalBossFight).

This is to include other dungeon specific functions which would disrupt the structure with unnecessary checks, such as combatEnd (allows the player to fight another enemy or prompt them to come back as there is no enemies left in the dungeon) and flee (if successful, allows the player to escape and return to the crossroads).

Also fixed small issues with interactivity, such as the player's Max health not changing when purchasing a health upgrade and being taken back to the town location after purchasing any upgrades (instead of staying at the blacksmith/tavern)