User input safeguard & Enemy remove fix
During testing, it was noticeable that the player could break the program by inputting a word rather than a number due to each switch case only having integer values.

To fix this, the hashString converts specific strings into Enum index which checks the users input and returns the appropriate integer.

Also found during testing was an oversight on a function where defeated enemies were continuously being re-added to the vector instead of being removed. This is now fixed where an iterator checks for the defeated enemy's ID in the vector and removes it accordingly.

Finally, changed the cost of items and raised the gold rewards of some enemies for better balancing and implemented a isGameEnd boolean to allow the information from the gameEnd function to show rather than having visual bugs such as no text displaying due to the program thinking the player is still in the dungeon/castle.