Game end fix and system clear oversights
Fixed an issue where after fighting the boss the game over text did not show up and instead took the player back to the crossroads. This was fixed by adding the isGameEnd bool to the while loop on the crossroad function and adding a break in the first case of the switch in the castle function.

During testing it was found that there were more oversights where information that was no longer needed weren't being cleared from the screen due to an absence of the system cls command in specific sections. This has now been rectified.